Relaxation Runge Kutta -- Entropy implementation (#195)
This PR further develops the RRK implementation in PHiLiP. The major
changes are:
- Split RRK ode solver into Root-finding (new) and Algebraic (same as
was previously implemented) classes.
- RK (explicit and implicit) can use the matrix-free
`apply_inverse_global_mass_matrix/apply_global_mass_matrix`.
- Add Heun's method (explicit RK) and third-order diagonally-implicit
RK; updated 2nd order DIRK as the original one was a placeholder.
- RRK classes no longer are derived from RungeKuttaODESolver. Rather,
they are derived from an EmptyRRKBase, and RungeKuttaODESolver has an
EmptyRRKBase type object. It does not do anything unless the user
selects an RRK type.
- Fixed issue #252 . As a result, the CPLUS scaling studies fail because
p is outside the allowable range of 2 to 5. I have changed them to use
cHu, which passes and still has the same intention as the original tests
(that is, FR does not add run-time computational cost).
- Added tests to verify that RRK works as intended.

This PR will close issue #174 and issue #252.

---------

Co-authored-by: Julien Brillon <43619715+jbrillon@users.noreply.github.com>
Co-authored-by: AlexanderCicchino <alexander.cicchino@mail.mcgill.ca>
Co-authored-by: A. Cicchino <39313421+AlexanderCicchino@users.noreply.github.com>